functions/http_file_upload/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
google-api-core==1.26.1
google-auth==1.28.0
google-cloud-core==1.6.0
google-cloud-storage==1.16.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
idna==2.10
numpy==1.20.1
packaging==20.9
pandas==0.24.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
urllib3==1.26.4
